move dependencies to pyproject.toml
Delete requirements.txt and move dependencies to pyproject.toml

This is the recommended file format by tutorials like
https://packaging.python.org/en/latest/tutorials/packaging-projects/#
and
https://setuptools.pypa.io/en/latest/userguide/quickstart.html
though several others are supported.

There should be a setuptools way to tell it in pyproject.toml
to read dependencies from a requirements.txt file, but I haven't
gotten this to work.

Also, the use cases for requirements.txt and pyproject.toml's
project.dependencies are potentially different, see various
discussions on the internet or on python packaging tutorial